handle Payments and create Subscriptions

+++ src/Paymill/LlamaKisses/Controllers/applicationcontroller.php
@@ -2,10 +2,14 @@
 
 namespace LlamaKisses\Controllers;
 
+use Monolog\Logger;
+
 abstract class ApplicationController {
 
   private    $action;
   private    $twig;
+  private    $log;
+
 
   protected  $urlvalues;
 
@@ -13,6 +17,7 @@ abstract class ApplicationController {
     $this->action = $action;
     $this->urlvalues = $urlvalues;
     $this->twig = $twig;
+    $this->log = new Logger( 'LLAMA_KISSES::ApplicationController' );
   }
 
   public function ExecuteAction() {
@@ -20,9 +25,16 @@ abstract class ApplicationController {
   }
 
   protected function ReturnView( $model = array() ) {
+    $this->log->addInfo( "Current user: " . $_SESSION['current_user'] );
+
+    $model['current_user'] = $_SESSION['current_user'];
     $yield = $this->twig->render( $_GET['controller'].'/'.$this->action.'.html', $model );
+
     $template = $this->twig->loadTemplate( 'layouts/application.html' );
-    var_dump( $_SESSION['current_user'] );
-    echo $template->render( array( 'yield' => $yield, 'current_user' => $_SESSION['current_user'] ) );
+    if( array_key_exists( 'name', $model ) ) {
+      echo $template->render( array( 'yield' => $yield, 'current_user' => $_SESSION['current_user'], 'current_user_name' => $model['name'] ) );
+    } else {
+      echo $template->render( array( 'yield' => $yield, 'current_user' => $_SESSION['current_user'] ) );
+    }
   }
 }
